cleanup: move active internal listener to an extension
