+++ envoy/network/connection_handler.h
@@ -242,5 +242,70 @@ public:
 
 using ActiveUdpListenerFactoryPtr = std::unique_ptr<ActiveUdpListenerFactory>;
 
+/**
+ * Internal listener callbacks.
+ */
+class InternalListener : public ConnectionHandler::ActiveListener {
+public:
+  virtual ~InternalListener() override = default;
+
+  /**
+   * Called when a new connection is accepted.
+   * @param socket supplies the socket that is moved into the callee.
+   */
+  virtual void onAccept(ConnectionSocketPtr&& socket) PURE;
+};
+
+using InternalListenerPtr = std::unique_ptr<InternalListener>;
+using InternalListenerOptRef = OptRef<InternalListener>;
+
+/**
+ * The query interface of the registered internal listener callbacks.
+ */
+class InternalListenerManager {
+public:
+  virtual ~InternalListenerManager() = default;
+
+  /**
+   * Return the internal listener binding the listener address.
+   *
+   * @param listen_address the internal address of the expected internal listener.
+   */
+  virtual InternalListenerOptRef
+  findByAddress(const Address::InstanceConstSharedPtr& listen_address) PURE;
+};
+
+using InternalListenerManagerOptRef =
+    absl::optional<std::reference_wrapper<InternalListenerManager>>;
+
+// The thread local registry.
+class LocalInternalListenerRegistry {
+public:
+  virtual ~LocalInternalListenerRegistry() = default;
+
+  // Set the internal listener manager which maintains life of internal listeners. Called by
+  // connection handler.
+  virtual void setInternalListenerManager(InternalListenerManager& internal_listener_manager) PURE;
+
+  // Get the internal listener manager to obtain a listener. Called by client connection factory.
+  virtual InternalListenerManagerOptRef getInternalListenerManager() PURE;
+
+  // Create a new active internal listener. Called by the server connection handler.
+  virtual InternalListenerPtr createActiveInternalListener(ConnectionHandler& conn_handler,
+                                                           ListenerConfig& config,
+                                                           Event::Dispatcher& dispatcher) PURE;
+};
+
+// The central internal listener registry interface providing the thread local accessor.
+class InternalListenerRegistry {
+public:
+  virtual ~InternalListenerRegistry() = default;
+
+  /**
+   * @return The thread local registry.
+   */
+  virtual LocalInternalListenerRegistry* getLocalRegistry() PURE;
+};
+
 } // namespace Network
 } // namespace Envoy
