+++ source/extensions/bootstrap/internal_listener/thread_local_registry.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "envoy/network/listener.h"
+#include "envoy/network/connection_handler.h"
 #include "envoy/thread_local/thread_local_object.h"
 
 namespace Envoy {
@@ -28,6 +28,11 @@ public:
     }
     return Network::InternalListenerManagerOptRef(*manager_);
   }
+
+  Network::InternalListenerPtr
+  createActiveInternalListener(Network::ConnectionHandler& conn_handler,
+                               Network::ListenerConfig& config,
+                               Event::Dispatcher& dispatcher) override;
 
 private:
   // The typical instance is the ``ConnectionHandlerImpl`` on the same thread.
