+++ source/server/connection_handler_impl.cc
@@ -10,7 +10,6 @@
 #include "source/common/network/address_impl.h"
 #include "source/common/network/utility.h"
 #include "source/common/runtime/runtime_features.h"
-#include "source/server/active_internal_listener.h"
 #include "source/server/active_tcp_listener.h"
 
 namespace Envoy {
@@ -53,12 +52,13 @@ void ConnectionHandlerImpl::addListener(absl::optional<uint64_t> overridden_list
     if (overridden_listener.has_value()) {
       if (auto iter = listener_map_by_tag_.find(overridden_listener.value());
           iter != listener_map_by_tag_.end()) {
-        iter->second->internalListener()->get().updateListenerConfig(config);
+        iter->second->internalListener()->updateListenerConfig(config);
         return;
       }
       IS_ENVOY_BUG("unexpected");
     }
-    auto internal_listener = std::make_unique<ActiveInternalListener>(*this, dispatcher(), config);
+    auto internal_listener =
+        local_registry->createActiveInternalListener(*this, config, dispatcher());
     details->typed_listener_ = *internal_listener;
     details->listener_ = std::move(internal_listener);
   } else if (config.listenSocketFactory().socketType() == Network::Socket::Type::Stream) {
@@ -132,7 +132,7 @@ void ConnectionHandlerImpl::addListener(absl::optional<uint64_t> overridden_list
         }
       }
     }
-  } else if (absl::holds_alternative<std::reference_wrapper<ActiveInternalListener>>(
+  } else if (absl::holds_alternative<std::reference_wrapper<Network::InternalListener>>(
                  details->typed_listener_)) {
     internal_listener_map_by_address_.insert_or_assign(
         config.listenSocketFactory().localAddress()->asStringView(), details);
@@ -276,10 +276,9 @@ ConnectionHandlerImpl::ActiveListenerDetails::udpListener() {
   return (val != nullptr) ? absl::make_optional(*val) : absl::nullopt;
 }
 
-ConnectionHandlerImpl::ActiveInternalListenerOptRef
-ConnectionHandlerImpl::ActiveListenerDetails::internalListener() {
-  auto* val = absl::get_if<std::reference_wrapper<ActiveInternalListener>>(&typed_listener_);
-  return (val != nullptr) ? absl::make_optional(*val) : absl::nullopt;
+Network::InternalListenerOptRef ConnectionHandlerImpl::ActiveListenerDetails::internalListener() {
+  auto* val = absl::get_if<std::reference_wrapper<Network::InternalListener>>(&typed_listener_);
+  return (val != nullptr) ? makeOptRef(val->get()) : absl::nullopt;
 }
 
 ConnectionHandlerImpl::ActiveListenerDetailsOptRef
