+++ test/server/connection_handler_test.cc
@@ -111,39 +111,6 @@ public:
       Network::UdpListenerWorkerRouterPtr listener_worker_router_;
     };
 
-    class MockInternalListenerRegistery : public Network::InternalListenerRegistry {
-    public:
-      Network::LocalInternalListenerRegistry* getLocalRegistry() override {
-        return &local_registry_;
-      }
-      // This registry does not depend on envoy thread local. Put it here because it should not be
-      // used in an integration test.
-      class MockLocalInternalListenerRegistry : public Network::LocalInternalListenerRegistry {
-      public:
-        void setInternalListenerManager(
-            Network::InternalListenerManager& internal_listener_manager) override {
-          manager_ = &internal_listener_manager;
-        }
-
-        Network::InternalListenerManagerOptRef getInternalListenerManager() override {
-          if (manager_ == nullptr) {
-            return Network::InternalListenerManagerOptRef();
-          }
-          return Network::InternalListenerManagerOptRef(*manager_);
-        }
-
-      private:
-        Network::InternalListenerManager* manager_{nullptr};
-      };
-      MockLocalInternalListenerRegistry local_registry_;
-    };
-    class InternalListenerConfigImpl : public Network::InternalListenerConfig {
-    public:
-      InternalListenerConfigImpl(MockInternalListenerRegistery& registry) : registry_(registry) {}
-      Network::InternalListenerRegistry& internalListenerRegistry() override { return registry_; }
-      MockInternalListenerRegistery& registry_;
-    };
-
     // Network::ListenerConfig
     Network::FilterChainManager& filterChainManager() override {
       return inline_filter_chain_manager_ == nullptr ? parent_.manager_
@@ -167,11 +134,7 @@ public:
     const std::string& name() const override { return name_; }
     Network::UdpListenerConfigOptRef udpListenerConfig() override { return *udp_listener_config_; }
     Network::InternalListenerConfigOptRef internalListenerConfig() override {
-      if (internal_listener_config_ == nullptr) {
-        return Network::InternalListenerConfigOptRef();
-      } else {
-        return *internal_listener_config_;
-      }
+      return Network::InternalListenerConfigOptRef();
     }
     envoy::config::core::v3::TrafficDirection direction() const override { return direction_; }
     void setDirection(envoy::config::core::v3::TrafficDirection direction) {
@@ -201,7 +164,6 @@ public:
     const std::chrono::milliseconds listener_filters_timeout_;
     const bool continue_on_listener_filters_timeout_;
     std::unique_ptr<UdpListenerConfigImpl> udp_listener_config_;
-    std::unique_ptr<InternalListenerConfigImpl> internal_listener_config_;
     Network::ConnectionBalancerSharedPtr connection_balancer_;
     BasicResourceLimitImpl open_connections_;
     const std::vector<AccessLog::InstanceSharedPtr> access_logs_;
@@ -307,22 +269,6 @@ public:
     return listeners_.back().get();
   }
 
-  TestListener* addInternalListener(
-      uint64_t tag, const std::string& name,
-      std::chrono::milliseconds listener_filters_timeout = std::chrono::milliseconds(15000),
-      bool continue_on_listener_filters_timeout = false,
-      std::shared_ptr<NiceMock<Network::MockFilterChainManager>> overridden_filter_chain_manager =
-          nullptr) {
-    listeners_.emplace_back(std::make_unique<TestListener>(
-        *this, tag, /*bind_to_port*/ false, /*hand_off_restored_destination_connections*/ false,
-        name, Network::Socket::Type::Stream, listener_filters_timeout,
-        continue_on_listener_filters_timeout, access_log_, overridden_filter_chain_manager,
-        ENVOY_TCP_BACKLOG_SIZE, nullptr));
-    listeners_.back()->internal_listener_config_ =
-        std::make_unique<TestListener::InternalListenerConfigImpl>(mock_listener_registry_);
-    return listeners_.back().get();
-  }
-
   void validateOriginalDst(Network::TcpListenerCallbacks** listener_callbacks,
                            TestListener* test_listener, Network::MockListener* listener) {
     Network::Address::InstanceConstSharedPtr normal_address(
@@ -366,7 +312,6 @@ public:
   Network::Address::InstanceConstSharedPtr local_address_{
       new Network::Address::Ipv4Instance("127.0.0.1", 10001)};
   NiceMock<Event::MockDispatcher> dispatcher_{"test"};
-  TestListener::MockInternalListenerRegistery mock_listener_registry_;
   std::list<TestListenerPtr> listeners_;
   std::unique_ptr<ConnectionHandlerImpl> handler_;
   NiceMock<Network::MockFilterChainManager> manager_;
@@ -2393,69 +2338,6 @@ TEST_F(ConnectionHandlerTest, ShutdownUdpListener) {
 
   ASSERT_TRUE(deleted_before_listener_)
       << "The read_filter_ should be deleted before the udp_listener_ is deleted.";
-}
-
-TEST_F(ConnectionHandlerTest, DisableInternalListener) {
-  InSequence s;
-  Network::Address::InstanceConstSharedPtr local_address{
-      new Network::Address::EnvoyInternalInstance("server_internal_address")};
-
-  TestListener* internal_listener =
-      addInternalListener(1, "test_internal_listener", std::chrono::milliseconds(), false, nullptr);
-  EXPECT_CALL(internal_listener->socket_factory_, localAddress())
-      .WillRepeatedly(ReturnRef(local_address));
-  handler_->addListener(absl::nullopt, *internal_listener, runtime_);
-  auto internal_listener_cb = handler_->findByAddress(local_address);
-  ASSERT_TRUE(internal_listener_cb.has_value());
-
-  handler_->disableListeners();
-  auto internal_listener_cb_disabled = handler_->findByAddress(local_address);
-  ASSERT_TRUE(internal_listener_cb_disabled.has_value());
-  ASSERT_EQ(&internal_listener_cb_disabled.value().get(), &internal_listener_cb.value().get());
-
-  handler_->enableListeners();
-  auto internal_listener_cb_enabled = handler_->findByAddress(local_address);
-  ASSERT_TRUE(internal_listener_cb_enabled.has_value());
-  ASSERT_EQ(&internal_listener_cb_enabled.value().get(), &internal_listener_cb.value().get());
-}
-
-TEST_F(ConnectionHandlerTest, InternalListenerInplaceUpdate) {
-  InSequence s;
-  uint64_t old_listener_tag = 1;
-  uint64_t new_listener_tag = 2;
-  Network::Address::InstanceConstSharedPtr local_address{
-      new Network::Address::EnvoyInternalInstance("server_internal_address")};
-
-  TestListener* internal_listener = addInternalListener(
-      old_listener_tag, "test_internal_listener", std::chrono::milliseconds(), false, nullptr);
-  EXPECT_CALL(internal_listener->socket_factory_, localAddress())
-      .WillRepeatedly(ReturnRef(local_address));
-  handler_->addListener(absl::nullopt, *internal_listener, runtime_);
-
-  ASSERT_NE(internal_listener, nullptr);
-
-  auto overridden_filter_chain_manager =
-      std::make_shared<NiceMock<Network::MockFilterChainManager>>();
-  TestListener* new_test_listener =
-      addInternalListener(new_listener_tag, "test_internal_listener", std::chrono::milliseconds(),
-                          false, overridden_filter_chain_manager);
-
-  handler_->addListener(old_listener_tag, *new_test_listener, runtime_);
-
-  Network::MockConnectionSocket* connection = new NiceMock<Network::MockConnectionSocket>();
-
-  auto internal_listener_cb = handler_->findByAddress(local_address);
-
-  EXPECT_CALL(manager_, findFilterChain(_)).Times(0);
-  EXPECT_CALL(*overridden_filter_chain_manager, findFilterChain(_)).WillOnce(Return(nullptr));
-  EXPECT_CALL(*access_log_, log(_, _, _, _));
-  internal_listener_cb.value().get().onAccept(Network::ConnectionSocketPtr{connection});
-  EXPECT_EQ(0UL, handler_->numConnections());
-
-  testing::MockFunction<void()> completion;
-  handler_->removeFilterChains(old_listener_tag, {}, completion.AsStdFunction());
-  EXPECT_CALL(completion, Call());
-  dispatcher_.clearDeferredDeleteList();
 }
 } // namespace
 } // namespace Server
